# services/core/api/openapi.yaml
openapi: 3.0.3
info:
  title: Cloud DevBox API
  description: |
    Cloud DevBox 是一个云端开发环境平台，提供完整的 RESTful API 用于管理开发环境、模板、部署、协作等功能。
    
    ## 认证方式
    
    API 支持两种认证方式：
    
    1. **JWT Token** - 用于 Web 应用和桌面客户端
       - 通过 `/api/v1/auth/login` 获取 token
       - 在请求头中添加 `Authorization: Bearer <token>`
    
    2. **API Key** - 用于程序化访问和 CI/CD 集成
       - 通过 Web 控制台或 API 创建 API Key
       - 在请求头中添加 `X-API-Key: <api_key>`
    
    ## 速率限制
    
    - 默认限制：1000 次/小时/用户
    - 超出限制返回 429 状态码
    - 响应头包含 `X-RateLimit-Remaining` 和 `X-RateLimit-Reset`
    
    ## 错误处理
    
    所有错误响应遵循统一格式：
    ```json
    {
      "code": "ERROR_CODE",
      "message": "Human readable message",
      "details": {},
      "request_id": "uuid",
      "timestamp": "2026-01-13T00:00:00Z"
    }
    ```
  version: 1.0.0
  contact:
    name: Cloud DevBox Team
    email: support@clouddevbox.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.clouddevbox.io/api/v1
    description: Production server
  - url: https://staging-api.clouddevbox.io/api/v1
    description: Staging server
  - url: http://localhost:8080/api/v1
    description: Local development
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /auth/login:
    post:
      tags: [Authentication]
      summary: 用户登录
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /environments:
    get:
      tags: [Environments]
      summary: 列出环境
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: status
          in: query
          schema:
            type: string
            enum: [creating, running, stopped, failed, suspended, archived]
        - name: template_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedEnvironments'
    post:
      tags: [Environments]
      summary: 创建环境
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEnvironmentRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Environment'
        '429':
          $ref: '#/components/responses/RateLimited'
  /environments/{id}:
    get:
      tags: [Environments]
      summary: 获取环境详情
      parameters:
        - $ref: '#/components/parameters/EnvironmentId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Environment'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Environments]
      summary: 更新环境
      parameters:
        - $ref: '#/components/parameters/EnvironmentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEnvironmentRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Environment'
    delete:
      tags: [Environments]
      summary: 删除环境
      parameters:
        - $ref: '#/components/parameters/EnvironmentId'
      responses:
        '204':
          description: 删除成功
  /environments/{id}/start:
    post:
      tags: [Environments]
      summary: 启动环境
      parameters:
        - $ref: '#/components/parameters/EnvironmentId'
      responses:
        '200':
          description: 启动成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Environment'
  /environments/{id}/stop:
    post:
      tags: [Environments]
      summary: 停止环境
      parameters:
        - $ref: '#/components/parameters/EnvironmentId'
      responses:
        '200':
          description: 停止成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Environment'
  /environments/{id}/restart:
    post:
      tags: [Environments]
      summary: 重启环境
      parameters:
        - $ref: '#/components/parameters/EnvironmentId'
      responses:
        '200':
          description: 重启成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Environment'
  /environments/{id}/ssh-config:
    get:
      tags: [Environments]
      summary: 获取 SSH 配置
      parameters:
        - $ref: '#/components/parameters/EnvironmentId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SSHConfig'
...
  /templates:
    get:
      tags: [Templates]
      summary: 列出模板
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTemplates'
    post:
      tags: [Templates]
      summary: 创建模板
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTemplateRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'
  /templates/{id}:
    get:
      tags: [Templates]
      summary: 获取模板详情
      parameters:
        - $ref: '#/components/parameters/TemplateId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'
    put:
      tags: [Templates]
      summary: 更新模板
      parameters:
        - $ref: '#/components/parameters/TemplateId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTemplateRequest'
      responses:
        '200':
          description: 更新成功
    delete:
      tags: [Templates]
      summary: 删除模板
      parameters:
        - $ref: '#/components/parameters/TemplateId'
      responses:
        '204':
          description: 删除成功

  # ============================================================================
  # Deployments
  # ============================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    EnvironmentId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    TemplateId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    Unauthorized:
      description: 未授权
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIError'
...
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIError'
...
    RateLimited:
      description: 请求过于频繁
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIError'
  schemas:
...
    APIError:
      type: object
      properties:
        code:
          type: string
          example: "BAD_REQUEST"
        message:
          type: string
          example: "Invalid request parameters"
        details:
          type: object
          additionalProperties:
            type: string
        request_id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
      required: [code, message, timestamp]
    Pagination:
      type: object
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total_items:
          type: integer
          format: int64
        total_pages:
          type: integer
        has_next:
          type: boolean
        has_prev:
          type: boolean

    # ============================================================================
    # Authentication Schemas
    # ============================================================================
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          default: "Bearer"
        expires_in:
          type: integer
          description: Token 过期时间（秒）
        user:
          $ref: '#/components/schemas/User'

    # ============================================================================
    # User Schemas
    # ============================================================================
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        avatar_url:
          type: string
          format: uri
        role:
          type: string
          enum: [developer, admin, org_admin]
        mfa_enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Environment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        template_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [creating, running, stopped, failed, suspended, archived]
        resources:
          $ref: '#/components/schemas/ResourceConfig'
        env_vars:
          type: object
          additionalProperties:
            type: string
        ports:
          type: array
          items:
            $ref: '#/components/schemas/PortMapping'
        preview_url:
          type: string
          format: uri
        ssh_host:
          type: string
        ssh_port:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        last_active_at:
          type: string
          format: date-time
    CreateEnvironmentRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
        template_id:
          type: string
          format: uuid
        resources:
          $ref: '#/components/schemas/ResourceConfig'
        env_vars:
          type: object
          additionalProperties:
            type: string
        ports:
          type: array
          items:
            $ref: '#/components/schemas/PortMapping'
        git_repo_url:
          type: string
          format: uri
      required: [name, template_id]
    UpdateEnvironmentRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        resources:
          $ref: '#/components/schemas/ResourceConfig'
        env_vars:
          type: object
          additionalProperties:
            type: string
    ResourceConfig:
      type: object
      properties:
        cpu:
          type: string
          example: "2"
          description: CPU 核心数
        memory:
          type: string
          example: "4Gi"
          description: 内存大小
        storage:
          type: string
          example: "20Gi"
          description: 存储大小
    PortMapping:
      type: object
      properties:
        container_port:
          type: integer
        protocol:
          type: string
          enum: [tcp, udp]
          default: tcp
        public:
          type: boolean
          default: false
    SSHConfig:
      type: object
      properties:
        host:
          type: string
        port:
          type: integer
        username:
          type: string
        private_key:
          type: string
        config_snippet:
          type: string
          description: SSH config 文件片段
...
    PaginatedEnvironments:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Environment'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # ============================================================================
    # Template Schemas
    # ============================================================================
    Template:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        icon_url:
          type: string
          format: uri
        default_resources:
          $ref: '#/components/schemas/ResourceConfig'
        dockerfile:
          type: string
        setup_commands:
          type: array
          items:
            type: string
        is_official:
          type: boolean
        is_public:
          type: boolean
        version:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateTemplateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        default_resources:
          $ref: '#/components/schemas/ResourceConfig'
        dockerfile:
          type: string
        setup_commands:
          type: array
          items:
            type: string
        is_public:
          type: boolean
          default: false
      required: [name, dockerfile]
    UpdateTemplateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        default_resources:
          $ref: '#/components/schemas/ResourceConfig'
        dockerfile:
          type: string
        setup_commands:
          type: array
          items:
            type: string
        is_public:
          type: boolean
    PaginatedTemplates:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Template'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # ============================================================================
    # Deployment Schemas
    # ============================================================================
...